Reorganize project context
