+++ crates/assistant/src/ambient_context.rs
@@ -1,0 +1,3 @@
+mod current_project;
+
+pub use current_project::*;
