+++ crates/assistant/src/assistant.rs
@@ -1,8 +1,8 @@
+mod ambient_context;
 pub mod assistant_panel;
 pub mod assistant_settings;
 mod codegen;
 mod completion_provider;
-mod project_info;
 mod prompts;
 mod saved_conversation;
 mod streaming_diff;
